Fix missing exception handler in NettyHttpClient and ChannelResourceFactory
